New HeartSlice problem and test

+++ dolfin_mech/TimeVaryingConstant.py
@@ -37,19 +37,21 @@ class TimeVaryingConstant():
         assert (type(val_ini) in (int, float, list, numpy.ndarray))
         if (type(val_ini) in (int, float)):
             assert (type(val_fin) in (int, float))
-            self.val_ini = numpy.array([val_ini])
-            self.val_fin = numpy.array([val_fin])
-            self.val_cur = numpy.array([val_ini])
-            self.val_old = numpy.array([val_ini])
+            self.val_ini = numpy.array([val_ini], dtype=float)
+            self.val_fin = numpy.array([val_fin], dtype=float)
+            self.val_cur = numpy.array([val_ini], dtype=float)
+            self.val_old = numpy.array([val_ini], dtype=float)
             self.set_value = self.set_value_sca
         elif (type(val_ini) in (list, numpy.ndarray)):
             assert (type(val_fin) in (list, numpy.ndarray))
-            self.val_ini = numpy.array(val_ini)
-            self.val_fin = numpy.array(val_fin)
-            self.val_cur = numpy.array(val_ini)
-            self.val_old = numpy.array(val_ini)
+            self.val_ini = numpy.array(val_ini, dtype=float)
+            self.val_fin = numpy.array(val_fin, dtype=float)
+            self.val_cur = numpy.array(val_ini, dtype=float)
+            self.val_old = numpy.array(val_ini, dtype=float)
             self.set_value = self.set_value_vec
         self.val = dolfin.Constant(val_ini)
+        # print ("ini", self.val_ini)
+        # print ("fin", self.val_fin)
 
 
 
@@ -66,6 +68,7 @@ class TimeVaryingConstant():
     def set_value_vec(self,
             val):
 
+        # print (val)
         self.val.assign(dolfin.Constant(val))
 
 
@@ -73,6 +76,11 @@ class TimeVaryingConstant():
     def set_value_at_t_step(self,
             t_step):
 
+        # print ("set_value_at_t_step")
+        # print ("t_step", t_step)
+        # print ("ini", self.val_ini)
+        # print ("fin", self.val_fin)
+        # print ("cur", self.val_ini * (1. - t_step) + self.val_fin * t_step)
         self.set_value(self.val_ini * (1. - t_step) + self.val_fin * t_step)
 
 
@@ -80,9 +88,18 @@ class TimeVaryingConstant():
     def set_dvalue_at_t_step(self,
             t_step):
 
+        # print ("set_dvalue_at_t_step")
         self.val_old[:] = self.val_cur[:]
+        # print ("old", self.val_old)
+        # print ("t_step", t_step)
+        # print ("ini", self.val_ini)
+        # print ("fin", self.val_fin)
         self.val_cur[:] = self.val_ini * (1. - t_step) + self.val_fin * t_step
+        # print ("cur", self.val_cur)
+        # print ("dvalue", self.val_cur - self.val_old)
+        # print (self.val.str(1))
         self.set_value(self.val_cur - self.val_old)
+        # print (self.val.str(1))
 
 
 
@@ -94,4 +111,7 @@ class TimeVaryingConstant():
 
     def homogenize(self):
 
+        # print ("homogenize")
+        # print (self.val.str(1))
         self.set_value(0*self.val_ini)
+        # print (self.val.str(1))
